+++ src/index.ts
@@ -102,6 +102,10 @@ require('dotenv').config()
       describe: 'Write the (filtered) OpenAPI file to an output file',
       type: 'string'
     })
+    .option('oaRename', {
+      describe: 'Change the OpenAPI title & Postman collection name',
+      type: 'string'
+    })
     .option('envFile', {
       describe: 'Path to the .env file to inject environment variables',
       type: 'string'
@@ -201,6 +205,7 @@ require('dotenv').config()
   const envFile = options?.envFile || '.env'
   const filterFile = options.filterFile
   const oaOutput = options.oaOutput || ''
+  const oaRename = options.oaRename || ''
 
   const portman = new Portman({
     ...options,
@@ -218,7 +223,8 @@ require('dotenv').config()
     postmanConfigPath,
     envFile,
     filterFile,
-    oaOutput
+    oaOutput,
+    oaRename
   })
 
   if (options.uploadOnly) {