+++ src/types/PortmanOptions.ts
@@ -26,5 +26,6 @@ export interface PortmanOptions {
   cliOptionsFile?: string
   oaLocal?: string
   oaUrl?: string
+  oaRename?: string
   init?: boolean
 }